When to Use UnrolledUtilities

The functions and types exported by this package tend to perform better than their counterparts from Base and Base.Iterators in the scenarios listed below. Additional examples and more precise measurements can be found in the automatically generated tables of benchmarks.

Outline:

When to Use Unrolled Functions

Long iterators

Iterators with elements of different types

Reduction operations with non-constant return types

Operations with more than 2 levels of recursion

When to Use StaticOneTo and StaticBitVector

Iterators of Ints from 1 to N

Long iterators of Bools that get modified across loop iterations

When to Use Unrolled Functions

Long iterators

map has an unstable return type for iterators with lengths greater than 32:

julia> Test.@inferred map(one, Tuple(1:31));
julia> Test.@inferred map(one, Tuple(1:32));ERROR: return type NTuple{32, Int64} does not match inferred return type Tuple
julia> Test.@inferred unrolled_map(one, Tuple(1:32));

getindex has an unstable return type for Core.Const slices of length N > 10 from iterators with lengths greater than N + 2:

julia> first_11(itr) = itr[1:11]first_11 (generic function with 1 method)
julia> Test.@inferred first_11(Tuple(1:13));
julia> Test.@inferred first_11(Tuple(1:14));ERROR: return type NTuple{11, Int64} does not match inferred return type Tuple{Vararg{Int64}}
julia> unrolled_first_11(itr) = unrolled_take(itr, Val(11))unrolled_first_11 (generic function with 1 method)
julia> Test.@inferred unrolled_first_11(Tuple(1:14));

For benchmarks that indicate performance improvements when using unrolled functions with long iterators, see Isolated Unrolled Functions

Iterators with elements of different types

in has an intermediate type instability that triggers allocations for nonuniform iterators:

julia> @allocated () in ((1, 2), (1, 2, 3))32
julia> @allocated unrolled_in((), ((1, 2), (1, 2, 3)))0

any, all, and foreach have intermediate type instabilities that trigger allocations for nonuniform iterators with lengths greater than 32:

julia> const nonuniform_itr_of_length_32 = (ntuple(Returns((1, 2)), 31)..., (1, 2, 3));
julia> const nonuniform_itr_of_length_33 = (ntuple(Returns((1, 2)), 32)..., (1, 2, 3));
julia> @allocated any(isempty, nonuniform_itr_of_length_32)0
julia> @allocated any(isempty, nonuniform_itr_of_length_33)1024
julia> @allocated unrolled_any(isempty, nonuniform_itr_of_length_33)0

getindex has an unstable return type for nonuniform iterators when given non-constant (i.e., not Core.Const) indices, which can lead to intermediate type instabilities that trigger allocations:

julia> function add_lengths(itr)
           length_sum = 0
           for n in 1:length(itr)
               length_sum += length(itr[n])
           end
       endadd_lengths (generic function with 1 method)
julia> @allocated add_lengths(((1, 2), (1, 2, 3)))64
julia> function unrolled_add_lengths(itr)
           length_sum = 0
           for n in 1:length(itr)
               length_sum += unrolled_applyat(length, n, itr)
           end
       endunrolled_add_lengths (generic function with 1 method)
julia> @allocated unrolled_add_lengths(((1, 2), (1, 2, 3)))0

Note

How can unrolled_applyat be stable if n isn't a Core.Const?

For the example of add_lengths, the compiler must infer the return type of itr[::Int64] before it can compile the call to length. Since this return type depends on the index n, the compiler needs to insert a runtime lookup into the method table that determines which method of length to call, length(::Tuple{Int64, Int64}) or length(::Tuple{Int64, Int64, Int64}), and this triggers allocations.

For the example of unrolled_add_lengths, the compiler instead infers the return types of itr[::Core.Const(1)], itr[::Core.Const(2)], and so on for every index into itr. Then, it compiles a call to length for each of these return types, and it inserts a runtime switch instruction that determines which result of length to return for a particular value of n. As long as length itself only returns one type (in this case, Int64), this ensures that unrolled_add_lengths has no intermediate type instabilities.

In other words, unrolled_applyat combines multiple methods for length and getindex into a single method, replacing the inefficient method table lookup that switches between them with a simpler switch instruction.

Tip

When should getindex be replaced with unrolled_applyat?

The specific example above could be simplified by using mapreduce, instead of using a for-loop in conjunction with unrolled_applyat:

julia> @allocated mapreduce(length, +, ((1, 2), (1, 2, 3)))0

However, there are often situations in which it is not possible to replace loops with function calls, like when those loops are parallelized over CPU or GPU threads. Moreover, CUDA is unable to compile any kernels with type instabilities that trigger allocations, so unrolled_applyat is required in order to parallelize over nonuniform iterators on GPUs.

For benchmarks that indicate performance improvements when using unrolled functions with nonuniform iterators, see Isolated Unrolled Functions and Nested Unrolled Functions

Reduction operations with non-constant return types

reduce and accumulate have unstable return types when the return type of op is not constant, but only for iterator lengths greater than 32:

julia> Test.@inferred reduce(tuple, Tuple(1:32));
julia> Test.@inferred reduce(tuple, Tuple(1:33));ERROR: return type Tuple{Tuple{Tuple{Tuple{Tuple{Tuple{Tuple{Tuple{Tuple{Tuple{Tuple{Tuple{Tuple{Tuple{Tuple{Tuple{Tuple{Tuple{Tuple{Tuple{Tuple{Tuple{Tuple{Tuple{Tuple{Tuple{Tuple{Tuple{Tuple{Tuple{Tuple{Tuple{Int64, Int64}, Int64}, Int64}, Int64}, Int64}, Int64}, Int64}, Int64}, Int64}, Int64}, Int64}, Int64}, Int64}, Int64}, Int64}, Int64}, Int64}, Int64}, Int64}, Int64}, Int64}, Int64}, Int64}, Int64}, Int64}, Int64}, Int64}, Int64}, Int64}, Int64}, Int64}, Int64} does not match inferred return type Union{Int64, Tuple{Any, Int64}}
julia> Test.@inferred unrolled_reduce(tuple, Tuple(1:33));

For benchmarks that indicate performance improvements when using unrolled functions with nonuniform reductions, see Isolated Unrolled Functions

Operations with more than 2 levels of recursion

All functions in Julia have a default "recursion limit" of 2; unless this limit is modified, it forces any function that recursively calls itself 2 or more times to have an unstable return type:

julia> recursive_length(itr) =
           eltype(itr) <: Tuple ? mapreduce(recursive_length, +, itr) : length(itr)recursive_length (generic function with 1 method)
julia> Test.@inferred recursive_length(((1, 2), (1, 2, 3)));
julia> Test.@inferred recursive_length((((1,), (2,)), (1, 2, 3)));ERROR: return type Int64 does not match inferred return type Any
julia> unrolled_recursive_length(itr) =
           eltype(itr) <: Tuple ?
           unrolled_mapreduce(unrolled_recursive_length, +, itr) : length(itr)unrolled_recursive_length (generic function with 1 method)
julia> Test.@inferred unrolled_recursive_length((((1,), (2,)), (1, 2, 3)));

Note

Is there any other way to avoid the default recursion limit?

The default recursion limit applies to all functions defined in Base and Base.Iterators, so those functions will have unstable return types for more than 2 levels of recursion, even when all user-defined functions passed to them have had their recursion limits disabled. It is also impossible to modify the recursion limits of functions defined in Base from external packages. This means that the only way to avoid the default recursion limit is to not use certain functions from Base, and instead to define alternatives without any recursion limits.

For benchmarks that indicate performance improvements when using unrolled functions with recursive operations, see Recursive Unrolled Functions

When to Use StaticOneTo and StaticBitVector

Iterators of Ints from 1 to N

UnrolledUtilities.StaticOneTo — Type

StaticOneTo(N)

A lazy and statically sized analogue of Base.OneTo(N).

This iterator can only store the integers from 1 to N, so its output_type_for_promotion is NoOutputType(). An efficient method is provided for unrolled_take, but no other unrolled functions can use StaticOneTos as output types.

If an iterator only contains the integers from 1 to N ≥ 0, it is possible to provide the compiler with the values in the iterator in addition to their types by using a StaticOneTo, as opposed to a Tuple or something similar. This can allow the compiler to fully optimize out code that depends on those values, essentially moving the code's execution from run time to compilation time:

julia> @code_llvm debuginfo=:none mapreduce(abs2, +, (1, 2, 3))define i64 @julia_mapreduce_29408([3 x i64]* nocapture noundef nonnull readonly align 8 dereferenceable(24) %0) #0 {
top:
  %1 = getelementptr inbounds [3 x i64], [3 x i64]* %0, i64 0, i64 0
  %2 = getelementptr inbounds [3 x i64], [3 x i64]* %0, i64 0, i64 1
  %3 = getelementptr inbounds [3 x i64], [3 x i64]* %0, i64 0, i64 2
  %unbox = load i64, i64* %1, align 8
  %4 = mul i64 %unbox, %unbox
  %unbox2 = load i64, i64* %2, align 8
  %5 = mul i64 %unbox2, %unbox2
  %6 = add i64 %5, %4
  %unbox4 = load i64, i64* %3, align 8
  %7 = mul i64 %unbox4, %unbox4
  %8 = add i64 %6, %7
  ret i64 %8
}
julia> @code_llvm debuginfo=:none mapreduce(abs2, +, StaticOneTo(3))define i64 @julia_mapreduce_29430() #0 {
top:
  ret i64 14
}

Standard library functions can sometimes take advantage of this optimization, but for most non-trivial operations it is necessary to use unrolled functions:

julia> @code_llvm debuginfo=:none mapreduce(log, +, StaticOneTo(3))define double @julia_mapreduce_29442() #0 {
top:
  %0 = call double @j__log_29444(double 2.000000e+00, {}* inttoptr (i64 140385746298248 to {}*))
  %1 = fadd double %0, 0.000000e+00
  %2 = call double @j__log_29444(double 3.000000e+00, {}* inttoptr (i64 140385746298248 to {}*))
  %3 = fadd double %1, %2
  ret double %3
}
julia> @code_llvm debuginfo=:none unrolled_mapreduce(log, +, StaticOneTo(3))define double @julia_unrolled_mapreduce_29449() #0 {
top:
  ret double 0x3FFCAB0BFA2A2002
}

For benchmarks that indicate performance improvements when using StaticOneTos, see Very Long Iterators.

Note

Can the compiler infer iterator values in other scenarios?

The compiler can usually infer the values of iterators that only contain singletons when they are accessed using Core.Const indices, but this is not possible for non-singletons (e.g., integers) unless some special type of iterator is used (e.g., a StaticOneTo).

Long iterators of Bools that get modified across loop iterations

UnrolledUtilities.StaticBitVector — Type

StaticBitVector{N, [U]}(f)
StaticBitVector{N, [U]}([bit])

A statically sized analogue of BitVector with Unsigned chunks of type U, which can be constructed using either a function f(n) or a constant bit. By default, U is set to UInt8 and bit is set to false.

This iterator can only store Bools, so its output_type_for_promotion is a ConditionalOutputType. Efficient implementations are provided for all unrolled functions, though the methods for unrolled_map and unrolled_accumulate only apply when the first item in the output is a Bool.

Loops in Julia often allocate memory when a value larger than 32 bytes in size is modified across loop iterations (regardless of whether the loops are unrolled or not). Since Bools are represented by bytes, this limits certain types of loops to modifying bitmasks of no more than 32 Bools in order to avoid allocations. Unlike an iterator of Bools, though, a StaticBitVector stores 8 bits in every byte, which makes it possible to modify up to 256 bits at a time in loops without any allocations:

julia> random_bit_flips(itr) = reduce(
           (itr′, i) -> Base.setindex(itr′, !itr′[rand(1:i)], i),
           1:length(itr);
           init = itr,
       )random_bit_flips (generic function with 1 method)
julia> @allocated random_bit_flips(ntuple(Returns(true), Val(32)))0
julia> @allocated random_bit_flips(ntuple(Returns(true), Val(33)))4224
julia> @allocated random_bit_flips(StaticBitVector{256}(true))0

As with StaticOneTos, standard library functions can occasionally optimize StaticBitVectors as well as unrolled functions, but most complex use cases require unrolled functions.

For benchmarks that indicate performance improvements when using long StaticBitVectors that get modified across loop iterations, see Nested Unrolled Closures.